requirements.txt
openai
pandas
pytest-asyncio